requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
faster-whisper==1.0.2
